refactor: merge Sessions into Runs — single live view
Runs page now reads from SQLite via /api/sessions. Detail page shows
live message feed with SSE streaming. No separate Sessions nav/pages.
Removed apps/studio/frontend/app/sessions/.
